fix: mise script now working correctly
Working on arm support

+++ post_setup.sh
@@ -1,6 +1,7 @@
 #!/bin/bash
 
 os=`uname -s`
+arch=`uname -m`
 
 exists() {
   if [ $# -ne 1 ]; then
@@ -44,14 +45,24 @@ i() {
   fi
 }
 
-islinux() {
-  if [ $os = "Linux" ]; then
-    # https://github.com/pyenv/pyenv/wiki#suggested-build-environment
-    sudo apt install build-essential libssl-dev zlib1g-dev build-essential \
-      libbz2-dev libreadline-dev libsqlite3-dev curl \
-      libncursesw5-dev xz-utils tk-dev libxml2-dev libxmlsec1-dev libffi-dev liblzma-dev
+isarm() {
+  if [ "$arch" == "arm64" -o "$arch" == "aarch64" ]; then
     return 0
   fi
+  return 1
+}
+
+islinux64() {
+  if [ "$arch" == "x86_64" ]; then
+    if [ $os = "Linux" -a exists apt ]; then
+      # https://github.com/pyenv/pyenv/wiki#suggested-build-environment
+      sudo apt install build-essential libssl-dev zlib1g-dev build-essential \
+        libbz2-dev libreadline-dev libsqlite3-dev curl \
+        libncursesw5-dev xz-utils tk-dev libxml2-dev libxmlsec1-dev libffi-dev liblzma-dev
+      return 0
+    fi
+  fi
+
   return 1
 }
 
@@ -72,18 +83,73 @@ mise-install() {
   mise global "$1@latest"
 }
 
+havebrew() {
+  if ! command -v brew &> /dev/null; then
+    return 1
+  fi
+  return 0
+}
+
+install_brew_from_source() {
+
+  if command -v apt-get &> /dev/null; then
+    sudo apt-get install -y build-essential procps curl file git
+  elif command -v yum &> /dev/null; then
+    packageList="procps-ng curl file git perl-IPC-Cmd"
+
+    sudo yum groupinstall -y 'Development Tools'
+
+    for packageName in $packageList; do
+      rpm --quiet --query $packageName || sudo yum install -y $packageName
+    done
+
+  elif command -v pacman &> /dev/null; then
+    sudo pacman -S base-devel procps-ng curl file git
+  fi
+
+  if ! command -v ruby &> /dev/null; then
+    echo "You need to install ruby :-)"
+    return 1
+  fi
+
+  if [ ! -d $HOME/github ]; then
+    mkdir -p $HOME/github
+  fi
+
+  if [ ! -d $HOME/github/brew ]; then
+    git clone https://github.com/Homebrew/brew.git $HOME/github/brew
+
+    cd $HOME/github/brew
+
+    # Get new tags from remote
+    git fetch --tags
+
+    # Get latest tag name
+    latestTag=$(git describe --tags "$(git rev-list --tags --max-count=1)")
+
+    # Checkout latest tag
+    git checkout $latestTag -b "latest-tag-$latestTag"
+  fi
+
+  export PATH="$PATH:$HOME/github/brew/bin"
+  mise install ruby@3.3
+}
+
 install_brew_if_needed() {
-  if [ islinux -o ismac ]; then
-    if ! command -v brew &> /dev/null; then
-      echo "brew not found, installing..."
+  if havebrew; then
+    return 0
+  fi
+
+  echo "brew not found, installing..."
+
+  if islinux64 || ismac; then
       sleep 1
       /bin/bash -c "$(curl -fsSL https://raw.githubusercontent.com/Homebrew/install/HEAD/install.sh)"
       # Run these two commands in your terminal to add Homebrew to your PATH:
       # #(echo; echo 'eval "$(/home/linuxbrew/.linuxbrew/bin/brew shellenv)"') >> $HOME/.zsh-homerc
       echo "Run this command to make brew immediately available"
-    fi
 
-    if islinux; then
+    if islinux64; then
       echo "PATH=/home/linuxbrew/.linuxbrew/bin:\$PATH" >> ~/.zsh-homerc
       eval "$(/home/linuxbrew/.linuxbrew/bin/brew shellenv)"
     elif ismac; then
@@ -96,16 +162,21 @@ install_brew_if_needed() {
 
     echo "export PATH" >> ~/.zsh-homerc
     source ~/.zsh-homerc
-  else
-    return 1
+    return 0
   fi
 
-  return 0
+  if isarm; then
+    install_brew_from_source
+    return $?
+  fi
+
+  return 1
+
 }
 
 install_brew_if_needed
 
-if command -v brew &> /dev/null; then
+if havebrew; then
   brew tap mike-engel/jwt-cli
   brew install fx fzf gcc bat cmatrix coreutils gnupg gnu-sed gnu-tar jwt-cli
   if ismac; then
@@ -134,10 +205,7 @@ if ! command -v wget &> /dev/null; then
   fi
 fi
 
-if ! command -v mise &> /dev/null; then
-  brew install mise
-  export PATH=$PATH:$HOMEBREW_PREFIX/bin
-
+if install_mise; then
   # the following do not work on mac so just using native package managers
   # mise-install bat
   # mise-install hyperfine
